rename valideringReducer
