+++ web/src/frontend/src/nav-soknad/redux/valideringReducer.ts
